requirements/base.txt

Django==1.11.1
django-cors-headers==2.5.2
django-filter==1.0.2
django-rest-swagger==2.1.2
djangorestframework==3.6.3

whitenoise==3.3.0
dj-database-url==0.4.2
dj-email-url==0.0.10




